// src/types/api.d.ts
export interface components {
  schemas: {
    StatusResponse: {
      status: string;
    };
    ServiceStatusResponse: {
      status: string;
      contractVersion: string;
    };
    LoginRequest: {
      email: string;
      password: string;
    };
    LoginResponse: {
      token: string;
    };
    MessageResponse: {
      message: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}